first pass at simple encryption
This isn't intended to be a complete solution yet, but I wanted to give
you a proof of concept I banged out. The encryption is completely
transparent to the end user. To use it, simply pass the parameter of
`encrypted => true` and it's encrypted & decrypted for you.

Example usage:

``` Puppet
datacat_fragment { 'encryption test':
  target => $target,
  data   => {
    name => 'fred',
  },
  encrypted => true,
}
```